refactor: move letter editor related component to Letter/Editor
