+++ app/step/3/components/Letter/Editor/ToolBar.jsx
@@ -1,4 +1,5 @@
-import { Container } from "./Container"
+import { Container } from "../../common/Container"
+
 const EditorToolBar = (props) => {
     const isActive = props.selectedOption !== ""
     return (
